+++ config/server/MultierImpl.js
@@ -21,65 +21,30 @@ function guid() {
         s4() + "-" + s4() + s4() + s4();
 }
 
-function getInformation(path) {
-    var nOffset = Math.max(0, Math.max(path.lastIndexOf("\\"), path.lastIndexOf("/")));
-    var eOffset = path.lastIndexOf(".");
-    if (eOffset < 0) {
-        eOffset = path.length;
-    }
-    const mimeType = mime.lookup(path);
-
-    return {
-        destination: path.substring(0, nOffset),
-        encoding: null,
-        fieldname: null,
-        mimetype: mimeType,
-        extension: mime.extension(mimeType),
-        charsets: mime.charsets.lookup(mimeType),
-        filename: path.substring(nOffset > 0 ? nOffset + 1 : nOffset, eOffset),
-        originalname: null,
-        path: path,
-        size: 0
-    };
-}
+/**
+ * File {
+ *      id: string
+ *      lastModified: long
+ *      name: "12923379_1598600950465914_2390381053412947001_n.jpg"
+ *      size: 120823
+ *      type: "image/jpeg"
+ * }
+ * @param path
+ * @return {{destination: string, encoding: null, name: null, mimetype, extension, charsets: *, filename: string, originalname: null, path: *, size: number}}
+ */
 
 module.exports = (app, requestPath, tempFolder, fieldName) => {
     const storage = multer.diskStorage({
         destination: tempFolder, // Specifies upload location...
         filename: function (req, file, cb) {
-            /*
-            switch (file.mimetype) { // *Mimetype stores the file type, set extensions according to filetype
-                case "image/jpeg":
-                    ext = ".jpeg";
-                    break;
-                case "image/png":
-                    ext = ".png";
-                    break;
-                case "image/gif":
-                    ext = ".gif";
-                    break;
-            }
-             */
             const id = guid();
-            file.filename = id;
+            file.id = id;
             fs.writeFileSync(path.normalize(tempFolder + "/" + id + ".json"), JSON.stringify(file), "utf8");
             cb(null, id);
         }
     });
     const upload = multer({ storage: storage });
 
-    /**
-     {
-       destination: "/Users/kamilbukum/DEV/robe/robe-react-ui/temp",
-       encoding: "7bit",
-       fieldname: "files",
-       filename: "e10fd61b-9f84-dd9f-e034-cb507a0e7df3",
-       mimetype: "application/x-x509-ca-cert",
-       originalname: "hamburgsud.com.crt",
-       path: "/Users/kamilbukum/DEV/robe/robe-react-ui/temp/e10fd61b-9f84-dd9f-e034-cb507a0e7df3",
-       size: 2067
-     }
-     */
 
     app.put(new RegExp(escapeRegexp(requestPath) + ".*"), fieldName ? upload.array(fieldName) : upload.array(), (req, res) => {
         res.status(200).send(req.files); // You can send any response to the user here
@@ -112,14 +77,13 @@ module.exports = (app, requestPath, tempFolder, fieldName) => {
         if (file.path) {
             filePath = file.path;
         } else {
-            var fileId = (typeof file === "string") ? file : file.filename;
+            var fileId = (typeof file === "string") ? file : file.id;
             filePath = tempFolder + "/" + fileId;
         }
-        var information = getInformation(filePath);
         fs.unlinkSync(filePath);
         fs.unlinkSync(filePath + ".json");
         log(filePath + " deleted from disk");
-        return information;
+        return file;
     }
 
 
@@ -135,22 +99,15 @@ module.exports = (app, requestPath, tempFolder, fieldName) => {
         });
 
         request.on("end", () => {
-            var bodyJson;
-            try{
-                bodyJson = JSON.parse(body);
-            } catch (e) {
-                bodyJson = {
-                    filename: body
-                };
-            }
-            if (Object.prototype.toString.call(bodyJson) === "[object Array]") {
+            let json = JSON.parse(body);
+            if (Object.prototype.toString.call(json) === "[object Array]") {
                 var files = [];
-                for (var i = 0; i < bodyJson.length; i++) {
-                    files[i] = deleteFileAndReturnFilename(bodyJson[i]);
+                for (var i = 0; i < json.length; i++) {
+                    files[i] = deleteFileAndReturnFilename(json[i]);
                 }
                 res.status(200).send(files);
             } else {
-                res.status(200).send(deleteFileAndReturnFilename(bodyJson));
+                res.status(200).send(deleteFileAndReturnFilename(json));
             }
         });
        // You can send any response to the user here