+++ site/components/samples/inputs/FileUploadInputSample.jsx
@@ -3,8 +3,6 @@ import { ShallowComponent } from "robe-react-commons";
 import FileUploadInput from "robe-react-ui/lib/inputs/upload/FileUploadInput";
 import { Button } from "react-bootstrap";
 
-const dataId = "3ae7baac-4e1c-6b63-4c07-b28d5b80821f";
-
 const filesUrl = "http://localhost:3000/files";
 
 let props = {
@@ -33,13 +31,16 @@ export default class FileUploadInputSample extends ShallowComponent {
                 <FileUploadInput
                     ref="dropZoneUploadNew"
                     name="files"
-                    id={dataId}
                     display="thumbnail"
                     label="Dosya Seçimi"
                     remote={props}
                     value={["info_test.png"]}
                     onChange={this.onChange}
+                    placeHolder="Upload Files"
+                    browseButtonText="Browse..."
+                    multiple={true}
                 />
+                <br />
                 <Button onClick={this.__onSubmitNew}>Gönder</Button>
             </div>
         );