+++ src/faicon/index.js
@@ -1,2 +1,1 @@
-export { default, default as Icon } from "./Icon";
 export { default as FaIcon } from "./FaIcon";