+++ src/inputs/upload/FileUploadInput.jsx
@@ -1,16 +1,34 @@
 import React from "react";
-import { ShallowComponent, Assertions, Generator } from "robe-react-commons";
-import StackLayout from "../../layouts/StackLayout";
-import { Table, Panel, ButtonGroup, Button, Glyphicon } from "react-bootstrap";
+import { findDOMNode } from "react-dom";
+import { Generator, Objects, Maps } from "robe-react-commons";
+import {FormGroup, FormControl, ControlLabel, Glyphicon, Checkbox, OverlayTrigger, Tooltip } from "react-bootstrap";
+import ValidationComponent from "../../validation/ValidationComponent";
+import Files from "../../util/Files";
 import FileManager from "../../util/FileManager";
+import DragDropLayout from "../../layouts/DragDropLayout";
+import ThumbnailGroup from "../../layouts/ThumbnailGroup";
+import ThumbnailItem from "../../layouts/ThumbnailItem";
+import FaIcon from "../../faicon/FaIcon";
 import "./FileUploadInput.css";
-// https://github.com/felixrieseberg/React-Dropzone-Component
 
-const supportMultiple = (typeof document !== "undefined" && document && document.createElement) ?
-    "multiple" in document.createElement("input") :
-    true;
+const supportMultiple = (typeof document !== "undefined" && document && document.createElement) ? "multiple" in document.createElement("input") : true;
 
-export default class FileUploadInput extends ShallowComponent {
+/**
+ *
+ *  File {
+ *      id: string
+ *      lastModified: long
+ *      name: "12923379_1598600950465914_2390381053412947001_n.jpg"
+ *      size: 120823
+ *      type: "image/jpeg"
+ *  }
+ *  FileUploadInput is a component which provides upload files.
+ *  Does necessary validations, rendering of validation messages.
+ *  @export
+ *  @class FileUploadInput
+ *  @extends {ValidationComponent}
+ */
+export default class FileUploadInput extends ValidationComponent {
     /**
      * Properties of the component
      *
@@ -18,7 +36,11 @@ export default class FileUploadInput extends ShallowComponent {
      */
     static propTypes: Map = {
         /**
-         * It is field name of uploaded file
+         * Style map for the component.
+         */
+        style: React.PropTypes.object,
+        /**
+         * name use as input field name
          */
         name: React.PropTypes.string.isRequired,
         /**
@@ -28,462 +50,331 @@ export default class FileUploadInput extends ShallowComponent {
         /**
          * File Id List or File Name
          */
-        value: React.PropTypes.arrayOf(React.PropTypes.string),
-        /**
-         * describe Selection File Multi or Single
-         */
-        multiple: React.PropTypes.bool,
-        /**
-         * FileUploadInputStyle
-         */
-        containerStyle: React.PropTypes.object,
+        value: React.PropTypes.oneOfType([
+            React.PropTypes.string,
+            React.PropTypes.arrayOf(React.PropTypes.string)
+        ]),
         /**
          * onChangeEvent event for the component
          */
         onChange: React.PropTypes.func,
         /**
+         * describe Selection File Multi or Single
+         */
+        multiple: React.PropTypes.bool,
+        /**
+         * Style map for the component.
+         */
+        itemStyle: React.PropTypes.object,
+        /**
+         * Item show renderer
+         */
+        itemRenderer: React.PropTypes.func,
+        /**
+         * File Input Form Control Place Holder
+         */
+        placeHolder: FormControl.propTypes.placeHolder,
+        /**
+         * it specifies that an input field is hidden or visible
+         */
+        hidden: React.PropTypes.bool,
+        /**
+         * auto upload is false then file will upload when clicking the upload button.
+         */
+        autoUpload: React.PropTypes.bool,
+        remote: React.PropTypes.object,
+        /**
          *
          */
-        onError: React.PropTypes.func,
-        /**
-         * Presentation Mode
-         */
-        display: React.PropTypes.oneOf(["list", "thumbnail"]),
+        toolbarPosition: React.PropTypes.string,
+        itemToolbarPosition: React.PropTypes.string,
         /**
          * Max Uploaded file
          */
         maxFileSize: React.PropTypes.number,
         /**
-         * auto upload is false then file will upload when clicking the upload button.
+         *Defines the display style of the Validation message.
          */
-        autoUpload: React.PropTypes.bool,
-        request: React.PropTypes.object,
-        /**
-         * will disable browse button
-         */
-        disabledBrowse: React.PropTypes.bool,
-        /**
-        *Defines the display style of the Validation message.
-        */
         validationDisplay: React.PropTypes.oneOf(['overlay', 'block'])
     };
 
-    /**
-     * defaultProps
-     * @static
-     */
-    static defaultProps = {
-        containerStyle: {
-            height: 400
-        },
-        display: "list",
+    static defaultProps: Map = {
         multiple: true,
-        autoUpload: true,
-        disabledBrowse: false,
+        itemStyle: {},
+        placeHolder: "Drag file to here",
         maxFileSize: 1000,
+        toolbarPosition: "bottom",
+        itemToolbarPosition: "bottom",
+        autoUpload: true,
         validationDisplay: "block"
     };
 
-    /**
-     * @type {Array}
-     * @access private
-     */
-
+    __maxSizeOfWillUploaded;
+    __fileInputEl;
+    __style;
+    __itemRenderer;
     __fileManager;
-
-    // current items
-    __value = [];
-    /**
-     * Example File : {
-	 *      destination: "/Users/kamilbukum/DEV/robe/robe-react-ui/temp",
-	 *      encoding: "7bit",
-	 *      fieldname: "files",
-	 *      filename: "e10fd61b-9f84-dd9f-e034-cb507a0e7df3",
-	 *      mimetype: "application/x-x509-ca-cert",
-	 *      originalname: "examplefile.png",
-	 *      path: "/Users/kamilbukum/DEV/robe/robe-react-ui/temp/e10fd61b-9f84-dd9f-e034-cb507a0e7df3",
-	 *      size: 2067
-     * }
-     * current files
-     */
-    __files = [];
-    __uploadedFiles = [];
-    __maxFileSize = 0;
+    __dragDropLayoutDom;
+    __value;
+    __initialValue;
     constructor(props) {
         super(props);
-        // init component
-        this.__init(props);
+        this.__componentId = Generator.guid();
+        this.componentWillReceiveProps(props);
     }
 
-    /**
-     * @desc initialize input
-     */
-    __init(props) {
-        this.__maxFileSize = props.multiple ? props.maxFileSize : 1;
-        if (props.value) {
-            this.__value = props.multiple ? props.value : [props.value];
+    componentWillReceiveProps(nextProps){
+        this.__fileManager = new FileManager(nextProps.remote);
+        this.__itemRenderer = this.props.itemRenderer || this.itemRenderer;
+
+        if(this.props.multiple) {
+            this.__maxSizeOfWillUploaded = supportMultiple ? this.props.maxFileSize : 1;
+            this.__value = this.props.value ? this.props.value: [];
         } else {
-            this.__value = [];
-        }
-        this.__fileManager = new FileManager(this.props.remote);
-        this.state = {
-            value: []
-        };
-        this.__uploadedFiles = [];
-        if (this.__value.length > 0) {
-            this.__fileManager.info(this.__value, this.__onInitSuccess, (error) => {
-                this.__value = [];
-                this.__onError(error);
-            });
-        }
-    }
-    /**
-     * @desc success callback after loading given files.
-     * @param {Array} files
-     * @access private
-     */
-    __onInitSuccess(files) {
-        for (let i = 0; i < files.length; i++) {
-            let file = files[i];
-            let fileIndex = this.__value.indexOf(file.filename);
-            if (!file.filename || fileIndex === -1) {
-                this.__onError({
-                    message: `Gelen ${file.path} dosyasında filename kimliği bulunamadı ! `
-                });
-                return false;
+            this.props.maxFileSize = 1;
+            if(this.props.value) {
+                this.__value = [this.props.value];
             }
-            this.__files[fileIndex] = file;
         }
-
+        this.state = {
+            files: {},
+            upload: false
+        };
+        if(this.__value && this.__value.length > 0) {
+            this.__fileManager.info(this.__value, this.onLoad, this.onError);
+        }
+        this.__initialValue = this.__value;
+    }
+    onLoad(fileArray){
+        let files = {};
+        for(let i = 0 ; i < fileArray.length; i++) {
+            let file = fileArray[i];
+            file.key = file.id;
+            file.isUploaded = true;
+            files[file.key] = file;
+        }
         this.setState({
-            items: this.__value
+            files: files,
+            upload: false
         });
-        this.onChange("load", this.__value);
-        return true;
     }
 
+    render() {
+        let attributes = {
+            type: "file",
+            name: this.props.name,
+            style: {display: "none"},
+            multiple: this.props.multiple,
+            ref: el => this.__fileInputEl = findDOMNode(el),
+            onChange: this.onFileSelect
+        };
+
+        let label = (this.props.label === undefined) ? undefined : (
+                <ControlLabel> {this.props.label} </ControlLabel>
+            );
+
+        return (
+            <FormGroup hidden={this.props.hidden} >
+                {label}
+                <FormControl
+                    {...attributes}
+                />
+                { /* this.createToolbar("top") */ }
+                <DragDropLayout
+                    ref={(el) => { this.__dragDropLayoutDom = findDOMNode(el)}}
+                    onDrop={this.onDrop}
+                    onClick={this.browse}
+                >
+                    <ThumbnailGroup
+                        id={this.__componentId}
+                    >
+                        {this.renderItems(this.state.files)}
+                    </ThumbnailGroup>
+                </DragDropLayout>
+                { /* this.createToolbar("bottom") */ }
+            </FormGroup>
+        )
+    }
+
+    createToolbar(position){
+        if(this.props.toolbarPosition !== position) {
+            return null;
+        }
+        return (
+            <div className={`rb-upload-toolbar rb-radius-${this.props.toolbarPosition}`}>
+                <Checkbox className="pull-left" checked readOnly>
+                    Select All
+                </Checkbox>
+                <Glyphicon className="pull-right" glyph="cloud-download"/>
+                <Glyphicon className="pull-right" glyph="cloud-upload"/>
+                <Glyphicon className="pull-right" glyph="remove"/>
+            </div>
+        );
+    }
+
+    onFileSelect(e) {
+        const target = e.dataTransfer ? e.dataTransfer : e.target;
+        this.onDrop({
+            action: "browse",
+            target
+        });
+    }
+
+    onDrop(e) {
+        if(!e.target) return ;
+        const files = e.target.files;
+        if(!files || files.length === 0 ) return ;
+        // Files dropped.
+        let droppedFiles = Files.getDroppedFiles(files);
+
+        if(this.autoUpload) { // if autoUpload true then call upload and break method.
+            this.upload(droppedFiles);
+            return ;
+        }
+        // if autoUpload is false then add files and change state.
+        this.setState({
+            files: this.mergeFiles(droppedFiles)
+        });
+    }
+
+    mergeFiles(newFileArray){
+        let newFiles = {};
+        for(let i = 0 ; i < newFileArray.length; i++) {
+            let file = newFileArray[i];
+            newFiles[file.key] = file;
+        }
+        return Objects.mergeClone(this.state.files, newFiles);
+    }
+
+    uploadFile(file){
+        this.upload([file])
+    }
+    upload(droppedFiles){
+        let onUpload = this.onUpload.bind(undefined, droppedFiles);
+        this.__fileManager.upload(droppedFiles, onUpload, this.onError.bind("upload"));
+    }
+
+    onUpload(droppedFiles, uploadedFiles){
+        for(let i = 0 ; i < uploadedFiles.length; i++) {
+            let previousKey = droppedFiles[i].key;
+            let previousFile = this.state.files[previousKey];
+            let file = uploadedFiles[i];
+            file = Objects.mergeClone(file, previousFile);
+            file.key = file.id;
+            file.isUploaded = true;
+            delete this.state.files[previousFile.key];
+            this.state.files[file.key] = file;
+            this.__value[this.__value.length] = file.id
+        }
+        this.state.upload = false;
+        this.setState({
+            files: this.state.files,
+            upload: true
+        });
+    }
+
+    deleteFile(file){
+        if(file.isUploaded && this.__initialValue.indexOf(file.id) === -1) { // if file uploaded then delete it on server.
+            this.__fileManager.delete(file, this.onDelete, this.onError.bind("delete"))
+        } else {
+            this.onDelete(file);
+        }
+    }
+
+    onDelete(file){
+        delete this.state.files[file.key];
+        let index = this.__value.indexOf(file.id);
+        if(index !== -1) { // if value is exist then this file deleted on server side.
+            delete this.__value[file.key];
+        }
+        this.state.isDeleted = false;
+        this.state.deletedFile = null;
+        this.setState({
+            isDeleted: true,
+            deletedFile: file
+        });
+    }
     /**
      * @desc Displays File Explorer to select file or files.
      * @access public
      */
-    browse() {
-        if (!this.props.disabledBrowse) {
-            this.open();
+    browse(e) {
+        let id = e.target.id;
+        let isClickLayout = id && (id.indexOf(this.__componentId) !== -1 || id === this.__dragDropLayoutDom.id);
+        if(isClickLayout) {
+            if (!this.props.disabledBrowse) {
+                this.open();
+            }
         }
     }
-
     /**
      * @desc Opens File Expoler to select file or files
      * @access private
      */
     open() {
-        this.fileInputEl.value = null;
-        this.fileInputEl.click();
+        this.__fileInputEl.value = null;
+        this.__fileInputEl.click();
     }
 
-    /**
-     * @desc Uploads files array by Using FileManager
-     * @param {Array} files
-     * @returns {boolean}
-     * @access private
-     */
-    upload(files: Array) {
-        if ((files.length + this.__value.length) > this.__maxFileSize) {
-            this.__onError({
-                message: `You cannot add more than ${this.__maxFileSize}`
-            });
-            return false;
-        }
-        this.__fileManager.upload(this.props.name, files, this.__onUploadSucess, this.__onError);
-        return true;
-    }
-
-    /**
-     * @param {Array} uploadedFiles
-     * @return {boolean}
-     * @access private
-     */
-    __onUploadSucess(files: Array) {
-        if (!files || !Assertions.isArray(files) || files.length === 0) {
-            this.__onError({
-                message: "Upload Failed ! "
-            });
-            return false;
-        }
-        let oldValue = this.__value;
-        this.__value = oldValue.slice(0);
-        for (let i = 0; i < files.length; i++) {
-            let file = files[i];
-            let fileIndex = this.__value.indexOf(file.filename);
-            if (fileIndex !== -1) {
-                this.__onError({
-                    message: "File Name Exist ! "
-                });
-                return false;
-            }
-            this.__value.push(file.filename);
-            fileIndex = this.__value.indexOf(file.filename);
-            this.__files[fileIndex] = file;
-
-            // new file uploaded
-            this.__uploadedFiles.push(file.filename);
-        }
-        this.setState({
-            items: this.__value
-        });
-        this.onChange("upload", this.__value, oldValue);
-        return true;
-    }
-
-    /**
-     * catchs any FileUploadInput's errors
-     * @param {Map} error
-     * @access private
-     */
-    __onError(error: Map) {
-        if (this.props.onError) {
-            this.props.onError(error);
-        }
-    }
-
-    /**
-     *
-     * @param {Map} item
-     * @returns {boolean}
-     * @access public
-     */
-    deleteItem(item: Map) {
-        // if is uploaded yet then delete it from server.
-        if (this.__uploadedFiles.indexOf(item.filename) !== -1) {
-            this.__fileManager.delete(item, () => {
-                this.__onDelete(item);
-            }, (error) => {
-                error.key = item.filename;
-                this.__onError(error);
-            });
-            return true;
-        }
-        this.__onDelete(item);
-        return true;
-    }
-
-    /**
-     * @param {Map} item
-     * @access private
-     */
-    __onDelete(item: Map) {
-        // clone items
-        let oldValue = this.__value;
-        let newValue = oldValue.slice(0);
-
-        // find item by filename
-        let index = newValue.indexOf(item.filename);
-        // delete item
-        newValue.splice(index, 1);
-        this.__files.splice(index, 1);
-        this.setState({
-            items: newValue
-        });
-        this.onChange("delete", newValue, oldValue);
-    }
-
-    render() {
-        return (
-            <Panel>
-                <StackLayout
-                    style={this.props.containerStyle}
-                    display={this.props.display}
-                    label={this.props.label}
-                    items={this.__files}
-                    onItemRender={this.onItemRender}
-                    onDragStart={this.onDragStart}
-                    onDragEnter={this.onDragEnter}
-                    onDragOver={this.onDragOver}
-                    onDragLeave={this.onDragLeave}
-                    onDrop={this.onDrop}
-                    toolbar={this.createButtons()}
-                    toolbarPosition="bottom"
-                    />
-            </Panel>
-        );
-    }
-    /**
-     * @desc render items
-     * @param {Map} item
-     * @param {string} display
-     * @returns {Object}
-     */
-    onItemRender(item: Map, display) {
-        switch (display) {
-            case "thumbnail":
-                return this.thumbnail(item);
-            default:
-                return this.list(item);
-        }
-    }
-
-    /**
-     * @desc render item will shown in thumbnail layout
-     * @param {Map} item
-     * @returns {Array<Object>}
-     */
-    thumbnail(item: Map) {
-        let onItemDelete = this.deleteItem.bind(this, item);
-        let uploadButton = null;
-
-        if (this.__uploadedFiles.indexOf(item.filename) !== -1) {
-            uploadButton = (
-                <Button bsSize="xsmall">
-                    <Glyphicon glyph="upload" />
-                </Button>
+    renderItems(files){
+        let elements = [];
+        Maps.forEach(files, (file, key) => {
+            let deleteIcon =  <FaIcon code="fa-minus-square-o pull-right"  size="fa-sm" onClick={this.deleteFile.bind(undefined, file)} />
+            let uploadIcon = !file.isUploaded ? (
+                    <FaIcon code="fa-upload pull-left"  size="fa-sm" onClick={this.uploadFile.bind(undefined, file)} />
+                ): null;
+            elements.push(
+                <ThumbnailItem key={file.key} >
+                    <div className={`rb-thumbnail-toolbar pull-top`}>
+                        {uploadIcon}
+                        {deleteIcon}
+                    </div>
+                    <div className="rb-upload-input" style={{...this.__style,  ...this.props.itemStyle}}>
+                        { this.__itemRenderer(file)}
+                    </div>
+                </ThumbnailItem>
             );
+        });
+        return elements;
+    }
+
+    toolTip(file){
+        let sizeText = ` - ( ${ Files.presentSize(file.size)} )`;
+        return (
+            <Tooltip className="rb-upload-item-tooltip" id={this.__componentId + file.name}>
+                <label>
+                    {`${file.name}${sizeText}`}
+                </label>
+            </Tooltip>
+        );
+    }
+
+    itemRenderer(file){
+        let extension = Files.getExtensionByMime(file.type);
+        if(!extension) {
+            extension = Files.getExtensionByPath(file.name);
+            if(extension === "") extension = null;
         }
+        extension = extension ? extension.toUpperCase(): "FILE";
         return [
-            <span className="gi-5x">
-                <Glyphicon glyph="file" />
-            </span>,
-            <h5>{item.originalname}</h5>,
-            <p>{item.description}</p>,
-            <div
-                style={{
-                    margin: 5
-                }}
-                >
-                <Button bsSize="xsmall" onClick={onItemDelete} >
-                    <Glyphicon glyph="remove" />
-                </Button>
-                {uploadButton}
-                <Button bsSize="xsmall">
-                    <Glyphicon glyph="zoom-in" />
-                </Button>
-            </div>
-        ];
+            <div className="rb-upload-image">
+                <div className="rb-upload-extension" >{extension}</div>
+                <div className="rb-upload-icon">
+                    <FaIcon code="fa-file-o"  />
+                </div>
+            </div>,
+            <OverlayTrigger placement="bottom" overlay={this.toolTip(file)}>
+                <div className="rb-upload-name">
+                    <div>
+                        {file.name}
+                    </div>
+                </div>
+            </OverlayTrigger>
+        ]
     }
 
-    /**
-     * @desc render item will shown in list layout
-     * @param item
-     * @returns {Object}
-     */
-    list(item: Map) {
-        return (
-            <Table
-                width="100%"
-                style={{
-                    display: "inline-block"
-                }}
-                >
-                <tbody>
-                    <tr>
-                        <td>
-                            <Glyphicon glyph="file" />
-                        </td>
-                        <td>{item.originalname}</td>
-                        <td>{item.mime}</td>
-                        <td>{item.size}</td>
-                    </tr>
-                </tbody>
-            </Table>
-        );
-    }
+    onError(action, error){
 
-    /**
-     *
-     * @returns {XML}
-     */
-    createButtons() {
-        const inputAttributes = {
-            type: "file",
-            style: { display: "none" },
-            multiple: supportMultiple && this.props.multiple,
-            ref: el => this.fileInputEl = el, // eslint-disable-line
-            onChange: this.onDrop
-        };
-        if (this.props.name && this.props.name.length) {
-            inputAttributes.name = this.props.name;
-        }
-
-
-        return (
-            <ButtonGroup className="pull-right">
-                <Button className="btn-file" bsSize="small" disabled={this.props.disabledBrowse} onClick={this.browse}>
-                    <Glyphicon glyph="folder-open" />&nbsp; <span className="hidden-xs">Browse …</span>
-                    <input
-                        {...inputAttributes}
-                        />
-                </Button>
-            </ButtonGroup>
-        );
-    }
-
-    /**
-     *
-     * @example File
-     * File = {
-     *       lastModified: 1468913662000,
-     *       lastModifiedDate: "Tue Jul 19 2016 10:34:22 GMT+0300 (EEST)",
-     *       name: "hamburgsud.com.crt",
-     *       size: 2067,
-     *       type: "application/x-x509-ca-cert",
-     *       webkitRelativePath: "",
-     *   }
-     * @desc called when any file drop on the presentation layout.
-     * @param {Object} e
-     * @returns {boolean}
-     */
-    onDrop(e) {
-        const droppedFiles = e.dataTransfer ? e.dataTransfer.files : e.target.files;
-        if (!droppedFiles || droppedFiles.length === 0) {
-            this.__onError({
-                message: "File couldn't droppped. You must select valid file."
-            });
-            return false;
-        }
-        for (let i = 0; i < droppedFiles.length; i++) {
-            droppedFiles[i].filename = Generator.guid();
-        }
-
-        if (this.props.autoUpload) {
-            this.upload(droppedFiles);
-        }
-        return true;
-    }
-
-    /**
-     * @desc called when added or removed files from FileUploadInput
-     * @param event
-     * @param newValue
-     * @param oldValue
-     */
-    onChange(event, newValue, oldValue) {
-        if (this.props.onChange) {
-            let e = {
-                target: {
-                    event: event,
-                    oldValue: this.__getValueByMultipleProperty(oldValue),
-                    value: this.__getValueByMultipleProperty(newValue),
-                    name: this.props.name
-                }
-            };
-            this.props.onChange(e);
-        }
-    }
-
-    __getValueByMultipleProperty(value: Array) {
-        if (!value || value.length === 0) {
-            return this.props.multiple ? [] : null;
-        }
-        return this.props.multiple ? value : value[0];
-    }
-    /**
-     * @ desc Decides ant update is necessary for re-rendering.
-     * Compares old props and state objects with the newer ones without going deep.
-     * @param {Object} nextProps
-     * @param {Object} nextState
-     * @returns {boolean} "true" component shoud update ,"false" otherwise.
-     */
-    shouldComponentUpdate(nextProps: Object, nextState: Object): boolean {
-        return super.shouldComponentUpdate(nextProps, nextState) || (this.state.items.length !== nextState.items.length);
     }
 }