+++ src/util/index.js
@@ -1,1 +1,3 @@
-export { default, default as FileManager } from "./FileManager";
+export { default as FileManager } from "./FileManager";
+export { default as Files } from "./Files";
+export { default as CSS } from "./CSS";